% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{construct_adj_mat}
\alias{construct_adj_mat}
\title{Construct an item-pool adjacency matrix.}
\usage{
construct_adj_mat(admin)
}
\arguments{
\item{admin}{An administration matrix with one row per respondent and one
column per item. Non-zero entries indicate that an item has been
administered to a respondent (see \code{\link[=meow]{meow()}} for details of the matrix-based
simulation state). A logical matrix is also accepted.}
}
\value{
An item-item adjacency matrix of type \code{matrix}.
}
\description{
For an item pool with N items, this returns an N x N matrix. The diagonal
elements contain the number of times each item has been administered. The
off-diagonal element \eqn{(i, j)} contains the number of respondents who have
been administered both item \eqn{i} and item \eqn{j}. In general this function
is not called directly, but is instead called within \code{\link[=meow]{meow()}}. It is exposed
to aid users who are testing item selection functions they have written.
}
\examples{
admin <- matrix(c(1, 1, 0,
                  1, 0, 1), nrow = 2, byrow = TRUE)
construct_adj_mat(admin)

}
