% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-loaders.R
\name{data_default}
\alias{data_default}
\title{A default data generation function that simulates normally distributed respondent abilities and item difficulties}
\usage{
data_default(N_persons = 100, N_items = 50, data_seed = 242424)
}
\arguments{
\item{N_persons}{Number of respondents to simulate}

\item{N_items}{Number of items to simulate}

\item{data_seed}{A random seed for generating reproducible data. This seed is re-initialized at the end of the data generation process}
}
\value{
A list with three components: A dataframe of item response named \code{resp}, a vector of true respondent abilities named \code{theta_tru}, and a vector of true item difficulties named \code{diff_tru}
}
\description{
A default data generation function that simulates normally distributed respondent abilities and item difficulties
}
