% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-loaders.R
\name{data_existing}
\alias{data_existing}
\title{Load data from existing files}
\usage{
data_existing(
  resp_path = "data/sample-resp.csv",
  theta_path = "data/true-sample-theta.rds",
  diff_path = "data/true-sample-diff.rds"
)
}
\arguments{
\item{resp_path}{A file path to a long form .csv file. File should have three columns, \code{id} which contains a numeric respondent identifier, \code{item} which contains a numeric item identifier, and resp which contains an item response. Be sure the form of the item response comports with the parameter update functions you choose to use.}

\item{theta_path}{A file path to a .rds file that contains a vector of true theta vales.}

\item{diff_path}{A file path to a .rds file that contains a vector of true item difficulty vales.}
}
\value{
A list with three components: A dataframe of item response named \code{resp}, a vector of true respondent abilities named \code{theta_tru}, and a vector of true item difficulties named \code{diff_tru}
}
\description{
Load data from existing files
}
