% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-loaders.R
\name{data_sample}
\alias{data_sample}
\title{Load a sample 1PL data for testing purposes}
\usage{
data_sample(data_args = list())
}
\value{
A list with three components: A dataframe of item response named \code{resp}, a dataframe of true person parameters named \code{pers_tru}, and a dataframe of true item parameters named \code{item_tru}
}
\description{
\code{data_sample()} loads a small sample dataset for testing purposes. These data are the same data returned by calling \code{data_simple_1pl()} with the default arguments.
}
