% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{edge_weight_inverse}
\alias{edge_weight_inverse}
\alias{edge_weight_negative_log}
\alias{edge_weight_linear}
\alias{edge_weight_power}
\alias{edge_weight_exponential}
\title{Alternative edge weight functions for network-based item selection}
\usage{
edge_weight_inverse(adj_mat, alpha = 1)

edge_weight_negative_log(adj_mat, alpha = 1)

edge_weight_linear(adj_mat, max_co_responses = NULL)

edge_weight_power(adj_mat, beta = 0.5, alpha = 1)

edge_weight_exponential(adj_mat, lambda = 0.1, alpha = 1)
}
\arguments{
\item{adj_mat}{The adjacency matrix where entry i,j is the number of co-responses between items i and j}

\item{alpha}{Smoothing parameter for avoiding division by zero}

\item{max_co_responses}{Scaling factor for linear weighting}

\item{beta}{Exponent for power transformation}

\item{lambda}{Decay constant for exponential decay weighting}
}
\value{
A matrix of edge weights for use in distance calculations
}
\description{
These functions provide different approaches to calculating edge weights from the adjacency matrix.
}
