% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meow-package.R
\name{fetch_meow}
\alias{fetch_meow}
\title{Fetch the MEOW data}
\usage{
fetch_meow(
  uri = file.path("https://www.arcgis.com/sharing/rest",
    "content/items/903c3ae05b264c00a3b5e58a4561b7e6/data"),
  save_file = NULL
)
}
\arguments{
\item{uri}{character, the uri of the data source}

\item{save_file}{character, to save the file, if left empty (""), NA or NULL
then no data is saved to disk.}
}
\value{
sf object
}
\description{
Fetch the MEOW data
}
