% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/meow.R
\docType{methods}
\name{meow}
\alias{meow}
\title{Download and display a random picture of a cat.}
\usage{
meow()
}
\value{
Nothing returned. Invoked for its side effect of displaying a cat picture.
}
\description{
Downloads and plots a random cat image (jpg) from \url{http://thecatapi.com}.
}
\examples{
meow()
}
\author{
Alex Chubaty and Kiri Whan
}

