% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meow-package.R
\docType{package}
\name{meow-package}
\alias{meow-package}
\title{meow: Get Random Pictures of Cats}
\description{
Queries http://thecatapi.com for cute cat photos and displays them in yor R session. Meow!
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/achubaty/meow}
  \item Report bugs at \url{https://github.com/achubaty/meow/issues}
}

}
\author{
\strong{Maintainer}: Alex M Chubaty \email{alex.chubaty@gmail.com}

Authors:
\itemize{
  \item Kiri Whan \email{kiriwhan@gmail.com}
}

}
\keyword{internal}
