% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/puRRR.R
\docType{package}
\name{puRRR-package}
\alias{puRRR}
\alias{puRRR-package}
\alias{purrr-package}
\title{Get random pictures of cats. Meow!}
\description{
Queries \url{http://thecatapi.com} for cute cat photos and displays them in yor R session.

Bug reports: https://github.com/achubaty/puRRR/issues
}
\author{
Alex M. Chubaty \email{alex.chubaty@gmail.com}

Kiri Whan\email{kiriwhan@gmail.com}
}
\keyword{package}

