% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meow-package.R
\name{read_meow}
\alias{read_meow}
\title{Read the ecoregions}
\usage{
read_meow(filename = system.file("exdata/meow.geojson", package = "meow"))
}
\arguments{
\item{filename}{character, the name of file to read}
}
\value{
sf object
}
\description{
Read the ecoregions
}
\examples{
\dontrun{
  library(sf)
  x <- meow::read_meow()
  plot(x['Lat_Zone'])
}
}
