% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_data_1pl}
\alias{sample_data_1pl}
\title{Small sample 1PL data}
\format{
\subsection{\code{sample_data_1pl}}{

A list with named components: \code{resp}, \code{pers_tru}, \code{item_tru}
Each of these components is a dataframe.
}
}
\source{
{meow} R pacakge
}
\usage{
sample_data_1pl
}
\description{
Some easy to use data, primarily for function testing and development
}
\keyword{datasets}
