% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-items.R
\name{select_max_dist}
\alias{select_max_dist}
\title{Item selection by network distance criterion.}
\usage{
select_max_dist(pers, item, R, admin, adj_mat = NULL, n_candidates = 1)
}
\arguments{
\item{pers}{A data frame of current respondent ability estimates.}

\item{item}{A data frame of current item parameter estimates.}

\item{R}{A respondent-by-item matrix of potential responses.}

\item{admin}{An integer administration matrix; \code{0} indicates an item has not
been administered to a respondent. See \code{\link[=meow]{meow()}} for details.}

\item{adj_mat}{An item-item adjacency matrix. See \code{\link[=construct_adj_mat]{construct_adj_mat()}}.}

\item{n_candidates}{The number of farthest items to assemble into a candidate
pool before selecting the next item by maximum information. Allows users to
trade off network density against estimation efficiency.}
}
\value{
An updated administration matrix with the selected item marked for
each respondent.
}
\description{
Administers the item farthest in the item network from the items a respondent
has already answered, with edges weighted by the inverse of their entry in the
item-item adjacency matrix. Ties are broken using the maximum information
criterion.
}
\examples{
sim <- meow(select_max_dist, update_theta_mle, data_simple_1pl,
            data_args = list(N_persons = 10, N_items = 10), fix = "item")
nrow(sim$results)

}
