% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-items.R
\name{select_sequential}
\alias{select_sequential}
\title{Item selection function that delivers the next item by item id number, simulating a fixed test form.}
\usage{
select_sequential(pers, item, resp, resp_cur = NULL, adj_mat = NULL)
}
\arguments{
\item{pers}{A dataframe of current respondent ability estimates.}

\item{item}{A dataframe of current item parameter estimates.}

\item{resp}{A long-form dataframe of all potential pre-simulated item responses.}

\item{resp_cur}{A long-form dataframe of administered item responses.}

\item{adj_mat}{An item-item adjacency matrix, where each entry is the count of individuals who have respondent to both item i and item j. See documentation for \code{construct_adj_mat}}
}
\value{
A long-form dataframe of all previously administered item responses with the new responses from this iteration appended to the end.
}
\description{
This function just administers the next item in a form, with the within-person item ordering being governed by the ordering the rows in the \code{resp} dataframe.
}
