% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-parameters.R
\name{update_theta_mle}
\alias{update_theta_mle}
\title{Update person ability via maximum likelihood estimation.}
\usage{
update_theta_mle(pers, item, R, admin)
}
\arguments{
\item{pers}{A data frame of current respondent parameter estimates.}

\item{item}{A data frame of item parameter values.}

\item{R}{A respondent-by-item matrix of potential responses.}

\item{admin}{An integer administration matrix; non-zero entries indicate
administered items. See \code{\link[=meow]{meow()}} for details.}
}
\value{
A list with two entries: \code{pers}, a data frame with updated respondent
ability estimates, and \code{item}, the unchanged data frame of item parameters.
}
\description{
This update function treats item parameters as fixed and known and updates
person ability estimates after each iteration with a maximum likelihood
estimate based on a 2PL item response function.
}
\examples{
data <- data_simple_1pl(N_persons = 10, N_items = 10)
admin <- matrix(0L, 10, 10)
admin[, 1:5] <- 1L
R <- matrix(data$resp$resp, nrow = 10, byrow = TRUE)
upd <- update_theta_mle(data$pers_tru, data$item_tru, R, admin)
head(upd$pers)

}
