% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-framework.R
\name{construct_adj_mat}
\alias{construct_adj_mat}
\title{Constructs an item pool adjacency matrix. For an item pool with $N$ items, this is an $N \times N$ matrix. The diagonal elements contain the number of times an item has been exposed. The off-diagonal elements contain the number of times the pair of items has been exposed to the same respondent. In general, this function is never called directly, but instead called within \code{cat_simulation()} calls.}
\usage{
construct_adj_mat(resp_cur, theta_tru, diff_tru)
}
\arguments{
\item{resp_cur}{A long-form dataframe of observed item responses.}

\item{theta_tru}{A vector of true respondent abilities.}

\item{diff_tru}{A vector of true item difficulties.}
}
\value{
An adjacency matrix of type \code{matrix}.
}
\description{
Constructs an item pool adjacency matrix. For an item pool with $N$ items, this is an $N \times N$ matrix. The diagonal elements contain the number of times an item has been exposed. The off-diagonal elements contain the number of times the pair of items has been exposed to the same respondent. In general, this function is never called directly, but instead called within \code{cat_simulation()} calls.
}
