% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{construct_adj_mat}
\alias{construct_adj_mat}
\title{Constructs an item pool adjacency matrix.}
\usage{
construct_adj_mat(resp_cur, pers_tru, item_tru)
}
\arguments{
\item{resp_cur}{A long-form dataframe of observed item responses.}

\item{pers_tru}{A dataframe of true respondent abilities.}

\item{item_tru}{A dataframe of true item parameters.}
}
\value{
An adjacency matrix of type \code{matrix}.
}
\description{
For an item pool with N items, this is an NxN matrix. The diagonal elements contain the number of times an item has been exposed. The off-diagonal elements contain the number of times the pair of items has been exposed to the same respondent. In general, this function is never called directly, but instead called within \code{meow()} calls. That said, it is exposed to the user to aid with testing other functions they may write.
}
