% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/meow.R
\docType{package}
\name{meow-package}
\alias{meow-package}
\title{Get random pictures of cats. Meow!}
\description{
Queries \url{http://thecatapi.com} for cute cat photos and displays them in yor R session.

Bug reports: https://github.com/achubaty/meow/issues
}
\author{
Alex M. Chubaty \email{alex.chubaty@gmail.com}

Kiri Whan\email{kiriwhan@gmail.com}
}
\keyword{package}

