% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meow-package.R
\docType{package}
\name{meow-package}
\alias{meow-package}
\alias{meow}
\title{meow package}
\description{
Serves TNC Marine Ecoregions Of the World (MEOW).
}
\seealso{
[TNC's ArcHub contribution](https://hub.arcgis.com/datasets/903c3ae05b264c00a3b5e58a4561b7e6/about)
}
