% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-items.R
\name{select_max_dist_enhanced}
\alias{select_max_dist_enhanced}
\title{Enhanced network-based item selection with configurable edge weights}
\usage{
select_max_dist_enhanced(
  pers,
  item,
  resp,
  resp_cur = NULL,
  adj_mat = NULL,
  n_candidates = 1,
  edge_weight_fun = edge_weight_inverse,
  edge_weight_args = list()
)
}
\arguments{
\item{pers}{A dataframe of current respondent ability estimates.}

\item{item}{A dataframe of current item parameter estimates.}

\item{resp}{A long-form dataframe of all potential pre-simulated item responses.}

\item{resp_cur}{A long-form dataframe of administered item responses.}

\item{adj_mat}{An item-item adjacency matrix.}

\item{n_candidates}{Number of farthest items to consider before applying information criterion.}

\item{edge_weight_fun}{Function to calculate edge weights from adjacency matrix.}

\item{edge_weight_args}{Additional arguments for the edge weight function.}
}
\value{
A long-form dataframe of all previously administered item responses with the new responses from this iteration appended to the end.
}
\description{
This function extends \code{select_max_dist} with flexible edge weight calculations.
}
