% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-items.R
\name{select_random}
\alias{select_random}
\title{Item selection function that delivers an item an item drawn at random from the item bank to each respondent.}
\usage{
select_random(
  pers,
  item,
  resp,
  resp_cur = NULL,
  adj_mat = NULL,
  select_seed = NULL
)
}
\arguments{
\item{pers}{A dataframe of current respondent ability estimates.}

\item{item}{A dataframe of current item parameter estimates.}

\item{resp}{A long-form dataframe of all potential pre-simulated item responses.}

\item{resp_cur}{A long-form dataframe of administered item responses.}

\item{adj_mat}{An item-item adjacency matrix, where each entry is the count of individuals who have respondent to both item i and item j. See documentation for \code{construct_adj_mat}}

\item{select_seed}{A random seed used only for item selection. Cleared each time this function is run.}
}
\value{
A long-form dataframe of all previously administered item responses with the new responses from this iteration appended to the end.
}
\description{
Each respondent has their own next item drawn at random from the remaining items.
}
