% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-parameters.R
\name{update_maths_garden}
\alias{update_maths_garden}
\title{Elo-style updates of person and item parameters}
\usage{
update_maths_garden(pers, item, resp, K_theta = 0.1, K_b = 0.1)
}
\arguments{
\item{pers}{A dataframe of current respondent parameter estimates.}

\item{item}{A dataframe of current item parameter estimates.}

\item{resp}{A long-form dataframe of only observed item responses.}

\item{K_theta}{User supplied learning rate for person ability updates. Defaults to 0.1}

\item{K_b}{User supplied learning rate for item difficulty updates. Defaults to 0.1}
}
\value{
An list of three objects, two of which are updated from the function input: \code{pers} is a dataframe with updated respondent parameter estimates, \code{item} is the dataframe of updated item parameter estimates. \code{resp_cur} is the dataframe of observed item responses.
}
\description{
This update function updates both person and item parameters according to the approach from the paper "Computer adaptive practice of Maths ability using a new item response model for on the fly ability and difficulty estimation" (Klinkenberg, Straatemeier, and van der Maas, 2011). Learning rates are tunable using supplied \code{K_theta} and \code{K_b} arguments.
}
