% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-parameters.R
\name{update_theta_mle}
\alias{update_theta_mle}
\title{Updated person parameters based on MLE estimates}
\usage{
update_theta_mle(pers, item, resp)
}
\arguments{
\item{pers}{A dataframe of current respondent parameter estimates.}

\item{item}{A dataframe of item parameter values.}

\item{resp}{A long-form dataframe of only observed item responses.}
}
\value{
An list of three objects, only one of which is updated from the function input: \code{pers} is a dataframe with updated respondent parameter estimates, \code{item} is the dataframe of item parameter values. \code{resp_cur} is the dataframe of observed item responses.
}
\description{
This update function treats item parameters as fixed and known, updating person ability estimates after each iteration according to a maximum likelihood estimate based upon a 2PL item response function.
}
